% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{build_version_to_concept_map}
\alias{build_version_to_concept_map}
\title{Map version record ids to concept record ids}
\usage{
build_version_to_concept_map(records)
}
\arguments{
\item{records}{List of Zenodo records.}
}
\value{
Named character vector mapping record id to concept id.
}
\description{
Map version record ids to concept record ids
}
